fix for 64-bit chocolatey installs
on 64-bit we need to invoke the install in a 64-bit environment.

we can use the interpreter flag to the chef script resource to do this.

i had a dificult time beating powershell_script into submission.  invoking cmd
to invoke powershell is a little gross, but the full path to the 64-bit
cmd.exe also looks like it'll be less brittle than the powershell path that
had the version string embedded in it.