This fixes the CI tests (#114)
* Use Ruby 2.2

A bunch of gems have updated to require Ruby 2.2

I started to track them down, but it was getting deep into Chef's
dependencies, etc. and making it hard.

* Travis is annoying and selected Ruby 2.2.0

Apparently, I have to specify 2.2.x to get a version greater than 2.2.0.

* Fixed rubocop lint on Gemfile

It now checks the order of `gem` statements? I didn't know that was an
issue.