Use the official install.ps1 (#113)
* Use the official install.ps1

This uses the `remote_file` resource to fetch the `install.ps1` from
chocolatey.org

This should allow improvements from the "official" copy to be matched
in the chocolatey version.

Closes #111
Closes #112
Closes #109

* Fixed ChefSpec/Fauxhai warning

The warning said in the future that this would be fatal.

* Use the official install.ps1

This uses the `remote_file` resource to fetch the `install.ps1` from
chocolatey.org.  The url to fetch `install.ps1` from is configurable in
case someone wants to run entirely detached from the internet.

This should allow improvements from the "official" copy to be matched
in the chocolatey version.

Closes #111
Closes #112
Closes #109

* Put chocolatey.org install.ps1 inside cookbook

This is the *stock* https://chocolatey.org/install.ps1 script.